Sketchy version, still to make use friendly and document it
