+++ bots/.gitignore
@@ -1,0 +1,3 @@
+#Creating this in order to keep directory empty
+*
+!.gitignore
