+++ db/commands.sh
@@ -1,0 +1,1 @@
+mongod --dbpath mongo/data --port 27017 --storageEngine wiredTiger --auth
