+++ handler/.gitignore
@@ -1,0 +1,2 @@
+#Creating this in order to keep directory empty
+__pycache__
