+++ handler/README.md
@@ -1,0 +1,14 @@
+# Handler
+
+Here is where all the scripts to run the discord bot and the game environment are stored.
+
+## Files
+
+- `bot.log` - Logfile of the discord bot
+- `funcs.py` - Helping functions and functions connecting with game environment
+- `handler.log` - Logfile for game environment
+- `handler.py` - Game environment script
+- `main.py` - Contains code of the discord bot
+- `run.txt` - File to sync the discord bot with game env.
+- `settings.json` - Contains variables of our env.
+- `settings.py` - Loads variables of the env.
