+++ handler/bot.log
@@ -1,0 +1,48 @@
+
+Timestamp: - 2018-04-06 23:42:46 File bot : pyth.zip, submitted, compiled and run successfully! Sending log file...
+Timestamp: - 2018-04-07 20:50:38 File bot : pyth.zip, submitted, compiled and run successfully! Sending log file...
+Timestamp: - 2018-04-07 20:52:08 File bot : testpip.zip, submitted, compiled and run successfully! Sending log file...
+Timestamp: - 2018-04-07 21:06:50 File bot : pyth.zip, submitted, compiled and run successfully! Sending log file...
+Timestamp: - 2018-04-07 21:08:34 File bot : pyth.zip, submitted, compiled and run successfully! Sending log file...
+Timestamp: - 2018-04-07 21:30:58 File bot : pyth.zip, submitted, compiled and run successfully! Sending log file...
+Timestamp: - 2018-04-08 14:20:44 File bot : pyth.zip, submitted, compiled and run successfully! Sending log file...
+Timestamp: - 2018-04-08 15:10:12 File bot : pyth.zip, submitted, compiled and run successfully! Sending log file...
+Timestamp: - 2018-04-08 15:12:18 File bot : pyth.zip, submitted but encountered an error compiling/running! Sending log file...
+Timestamp: - 2018-04-08 15:17:08 File bot : pyth.zip, submitted but encountered an error compiling/running! Sending log file...
+Timestamp: - 2018-04-08 21:19:47 File bot : pyth.zip, submitted but encountered an error compiling/running! Sending log file...
+Timestamp: - 2018-04-08 21:25:15 File bot : pyth.zip, submitted but encountered an error compiling/running! Sending log file...
+Timestamp: - 2018-04-08 21:26:27 File bot : pyth.zip, submitted but encountered an error compiling/running! Sending log file...
+Timestamp: - 2018-04-08 21:29:49 File bot : pyth.zip, submitted but encountered an error compiling/running! Sending log file...
+Timestamp: - 2018-04-09 20:09:26 'NoneType' object has no attribute 'get'
+Timestamp: - 2018-04-09 20:24:48 update only works with $ operators
+Timestamp: - 2018-04-09 20:29:44 name 'user' is not defined
+Timestamp: - 2018-04-09 20:32:23 name 'user' is not defined
+Timestamp: - 2018-04-09 20:34:50 name 'user' is not defined
+Timestamp: - 2018-04-09 20:35:40 name 'user' is not defined
+Timestamp: - 2018-04-09 20:37:32 update only works with $ operators
+Timestamp: - 2018-04-09 20:42:08 
+Timestamp: - 2018-04-09 20:42:51 
+Timestamp: - 2018-04-09 20:43:55 File bot : pyth.zip, submitted, compiled and run successfully! Sending log file...
+Timestamp: - 2018-04-09 20:51:22 
+Timestamp: - 2018-04-09 20:51:22 Session is closed
+Timestamp: - 2018-04-09 20:54:36 
+Timestamp: - 2018-04-09 20:54:36 Session is closed
+Timestamp: - 2018-04-09 20:56:13 string index out of range
+Timestamp: - 2018-04-09 20:57:36 module 'settings' has no attribute 'delete_one'
+Timestamp: - 2018-04-09 20:58:58 File bot : pyth.zip, submitted, compiled and run successfully! Sending log file...
+Timestamp: - 2018-04-09 21:00:09 File bot : java.zip, submitted, compiled and run successfully! Sending log file...
+Timestamp: - 2018-04-09 21:01:07 
+Timestamp: - 2018-04-09 21:01:07 Session is closed
+Timestamp: - 2018-04-10 17:52:42 INTERNAL SERVER ERROR (status code: 500): 500: Internal Server Error
+Timestamp: - 2018-04-10 17:55:22 unsupported operand type(s) for +: 'dict' and 'str'
+Timestamp: - 2018-04-10 17:58:22 unsupported operand type(s) for +: 'dict' and 'str'
+Timestamp: - 2018-04-10 17:59:34 unsupported operand type(s) for +: 'dict' and 'str'
+Timestamp: - 2018-04-10 18:01:05 
+Timestamp: - 2018-04-10 18:17:41 
+Timestamp: - 2018-04-10 18:32:57 Can't convert 'dict' object to str implicitly
+Timestamp: - 2018-04-10 18:38:42 name 'g' is not defined
+Timestamp: - 2018-04-10 18:41:09 'NoneType' object has no attribute 'get'
+Timestamp: - 2018-04-10 18:46:11 'NoneType' object has no attribute 'get'
+Timestamp: - 2018-04-10 18:47:55 'NoneType' object has no attribute 'get'
+Timestamp: - 2018-04-10 18:54:08 Can't convert 'dict' object to str implicitly
+Timestamp: - 2018-04-10 18:56:59 Can't convert 'dict' object to str implicitly
