+++ handler/handler.log
@@ -1,0 +1,6 @@
+Timestamp: - 2018-04-06 23:42:43 Running : Splinter#4298
+Timestamp: - 2018-04-07 20:50:38 Running : Splinter#4298
+Timestamp: - 2018-04-07 20:52:08 Running : Splinter#4298
+Timestamp: - 2018-04-07 21:06:50 Running : Splinter#4298
+Timestamp: - 2018-04-07 21:08:34 Running : Splinter#4298
+Timestamp: - 2018-04-07 21:30:59 Running : Splinter#4298
