--- settings.json
@@ -1,20 +1,0 @@
-{
-	"compileOut": 245,
-	"runOut": 125,
-	"token": "",
-	"admins": [
-		"Splinter#4298",
-		"Frank Whoee#1338"
-	],
-	"timeSub": "APRIL 12 3:30PM GMT",
-	"infos": "season-2/infos.txt",
-	"log": "bot.log",
-	"server": "HaliteTournaments",
-	"brackets": "imgs/brackets.png",
-	"submit": false,
-	"season": "2",
-	"onTour": false,
-	"matches": {},
-	"specs":"season-2/specs.txt",
-	"engineLink":"https://github.com/HaliteTournaments/Tournament-Environment/tree/master/env/halite-env"
-}
